Refactor scripts based on feedback shared in the review
