+++ packages/scripts/utils/index.js
@@ -8,22 +8,16 @@ const {
 	spawnScript,
 } = require( './cli' );
 const {
+	hasBabelConfig,
+	hasJestConfig,
+} = require( './config' );
+const {
 	fromConfigRoot,
 	hasProjectFile,
 } = require( './file' );
-const { hasPackageProp } = require( './package' );
-
-const hasBabelConfig = () =>
-	hasProjectFile( '.babelrc' ) ||
-	hasProjectFile( 'babel.config.js' ) ||
-	hasPackageProp( 'babel' );
-
-const hasJestConfig = () =>
-	hasCliArg( '-c' ) ||
-	hasCliArg( '--config' ) ||
-	hasProjectFile( 'jest.config.js' ) ||
-	hasProjectFile( 'jest.config.json' ) ||
-	hasPackageProp( 'jest' );
+const {
+	hasPackageProp,
+} = require( './package' );
 
 module.exports = {
 	fromConfigRoot,
